Auto-delete XLSX when converting to SQL
To prevent having multiple copies of the data lying around after conversion. The original SQL/YAML files are not deleted, since they're more important to the project and are considered the authoritative version of the data, but the spreadsheet looks nicer and could lure someone into using an older version of the data after it's already been exported back to CCI and changes have been made there.